+++ src/main.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { AppShell } from "./app/AppShell";
 import { SessionProvider } from "./features/session/SessionContext";
+import { SocialProvider } from "./features/social/SocialContext";
 import { ControllerConfigProvider } from "./features/gamepad/ControllerConfigContext";
 import { applyStoredTheme } from "./features/theme/useTheme";
 import "./styles/global.css";
@@ -13,9 +14,11 @@ applyStoredTheme();
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
     <SessionProvider>
-      <ControllerConfigProvider>
-        <AppShell />
-      </ControllerConfigProvider>
+      <SocialProvider>
+        <ControllerConfigProvider>
+          <AppShell />
+        </ControllerConfigProvider>
+      </SocialProvider>
     </SessionProvider>
   </React.StrictMode>,
 );